adding cursor animation snapshot
